Move presentation bounded code out of request model
Follow up to https://github.com/wmde/FundraisingFrontend/pull/928
